+++ client/katatema-hot-reloadable-client.js
@@ -3,7 +3,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import UpdatableComponent from "../src/updatable-component";
 
-global.modan = { eventEmitter: new EventEmitter() };
+global.katatema = { eventEmitter: new EventEmitter() };
 
 function evalScript(script) {
   const module = { exports: {} };
@@ -11,9 +11,9 @@ function evalScript(script) {
   return module.exports;
 }
 
-const mod = evalScript(window.modanData.componentScript)
+const mod = evalScript(window.katatemaData.componentScript)
 const Component = mod.default || mod;
 ReactDOM.render(
-  <UpdatableComponent Component={Component} eventEmitter={global.modan.eventEmitter}/>,
+  <UpdatableComponent Component={Component} eventEmitter={global.katatema.eventEmitter}/>,
   document.getElementById("container")
 );