+++ src/render.js
@@ -5,20 +5,20 @@ import React from "react";
 import ReactDOMServer from "react-dom/server";
 
 export default function render(relativePagePath, { hotReloadable }) {
-  const path = `${process.cwd()}/.modan/server-bundles/${relativePagePath}`;
+  const path = `${process.cwd()}/.katatema/server-bundles/${relativePagePath}`;
   if (fs.existsSync(path)) {
     const mod = require(path);
     const Component = mod.default || mod;
     const innerHtml = ReactDOMServer.renderToString(<Component/>);
     const directoryName = hotReloadable ? "client-bundles" : "server-bundles";
-    const componentScript = fs.readFileSync(`.modan/${directoryName}/${relativePagePath}`, "utf-8");
-    const modanData = { componentScript };
+    const componentScript = fs.readFileSync(`.katatema/${directoryName}/${relativePagePath}`, "utf-8");
+    const katatemaData = { componentScript };
     const html = ReactDOMServer.renderToString(
       <Html
         headChildren={(global.Head || Head).rewind()}
         hotReloadable={hotReloadable}
         innerHtml={innerHtml}
-        modanData={modanData}
+        katatemaData={katatemaData}
       />
     );
     return `<!DOCTYPE html>\n${html}\n`;